Create custom useQueryParams hook.
